+++ package.json
@@ -16,6 +16,7 @@
     "@testing-library/user-event": "^12.1.10",
     "firebase": "^8.10.0",
     "moment": "^2.29.1",
+    "node-sass": "^5.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-qr-code": "^2.0.3",
