+++ src/components/components/GeneralButton/StyledButton.jsx
@@ -1,0 +1,6 @@
+import { Button } from '@mui/material';
+import { styled } from '@mui/material/styles';
+
+export const StyledButton = styled(Button)(({theme}) => ({
+    borderRadius: "50px", 
+}))
