+++ src/components/components/GeneralSelect/GeneralSelect.jsx
@@ -1,5 +1,5 @@
 import React from "react"; 
-import { StyledSelect } from "../styled/general";
+import { StyledSelect } from "./StyledSelect";
 import { Typography, Box } from '@mui/material';
 
 export const GeneralSelect = ({label, children, ...props}) => {
