+++ src/components/components/Topbar/style.scss
@@ -1,0 +1,6 @@
+.Topbar {
+    &__logo {
+        cursor: pointer; 
+        margin-left: 20px;
+    }
+}
